Send the IRC message using an asynchronous version of the IRC Hooky Lambda function
The way this works is that the synchronous version of this function (the
one that is triggered through API Gateway) publishes an SNS message
with the details related to sending an IRC message.

The asynchronous version of the Lambda function (the one that is
triggered via the SNS message) then takes the details and sends the
IRC message.

The reasoning for this is that there is (currently) a hard 10-second
limit on requests triggered via API Gateway and unfortunately sending
IRC messages takes longer than that.

See this PR for more details: https://github.com/marvinpinto/irc-hooky/pull/8