Fix for the empty payload bug
Ran into a situation where API Gateway would seemingly send an empty
payload into the Lambda function.

After shaving many AWS yaks (figuring out IAM roles, enabling logging on
all the things, etc) I tracked it down to the API Gateway Mapping
template.

Looking at the docs here:

http://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-mapping-template-reference.html#input-variable-reference

You will notice that one of the examples is:

```
{
  "name" : "$input.params('name')",
  "body" : $input.json('$')
}
```

You will also notice that that is *not* valid JSON. This was the clue
for me.

I had assumed all along that the mapping template had to be valid json,
which was why using a function like `json.dumps` works and makes
sense!

In the end, it turns out that what I needed for the payload value was
indeed `$input.json('$')`, and this is why you will notice that
handcrafted artisanal "json like" mapping template.